requirements.txt
flask
pandas
gunicorn
openpyxl
numpy
tensorflow
Pillow
